+++ common/package.json
@@ -1,5 +1,7 @@
 {
   "name": "common",
+  "private": true,
+  "version": "1.0.0",
   "type": "module",
   "devDependencies": {
     "@types/bun": "latest"
