+++ server/package.json
@@ -1,5 +1,6 @@
 {
-  "name": "course-mate",
+  "name": "server",
+  "private": true,
   "version": "1.0.0",
   "description": "",
   "main": "src/index.ts",
