+++ web/package.json
@@ -9,13 +9,13 @@
     "preview": "next start"
   },
   "dependencies": {
+    "common": "workspace:*",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@fontsource/roboto": "^5.0.13",
     "@hookform/resolvers": "^3.9.1",
     "@mui/icons-material": "^5.16.7",
     "@mui/material": "^5.15.20",
-    "common": "workspace:common",
     "devalue": "^5.1.1",
     "firebase": "^10.12.2",
     "framer-motion": "^11.3.23",
